Core/Cluster: Overhaul expansion algorithm

+++ src/map/map.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "common.hpp"
+#include "offset.hpp"
 
 #include <inttypes.h>
 #include <vector>
@@ -12,17 +12,16 @@ template <typename E>
 class Map
 {
 public:
-
     Map(int32_t x, int32_t y, uint32_t dx, uint32_t dy);
 
-    Cell<E>* addTo2D(int16_t x, int16_t y, E e);
-    Cell<E>* addTo(int16_t q, int16_t r, E e);
+    Cell<E>* addTo2D(int32_t x, int32_t y, E e);
+    Cell<E>* addTo(int32_t q, int32_t r, E e);
     Cell<E>* addTo(const Offset&& offset, E e);
 
-    Cell<E>* get(int16_t q, int16_t r);
+    Cell<E>* get(int32_t q, int32_t r);
     Cell<E>* get(const Offset& offset);
 
-    Cell<E>* getOrCreate(int16_t q, int16_t r, bool siblings = true);
+    Cell<E>* getOrCreate(int32_t q, int32_t r, bool siblings = true);
     Cell<E>* getOrCreate(const Offset& offset, bool siblings = true);
 
     std::vector<Cell<E>*> createSiblings(Cell<E>* cell);
